Module casper_types::account

Contains types and constants associated with user accounts.

Re-exports

pub use self::action_thresholds::ActionThresholds;

pub use self::associated_keys::AssociatedKeys;

Modules

action_thresholds

This module contains types and functions for managing action thresholds.

associated_keys

This module contains types and functions for working with keys associated with an account.

Structs

Account

Represents an Account in the global state.

AccountHash

A newtype wrapping an array which contains the raw bytes of the AccountHash, a hash of Public Key and Algorithm

TryFromSliceForAccountHashError

Associated error type of TryFrom<&[u8]> for AccountHash.

Weight

The weight attributed to a given AccountHash in an account’s associated keys.

Enums

ActionType

The various types of action which can be performed in the context of a given account.

AddKeyFailure

Errors that can occur while adding a new AccountHash to an account’s associated keys map.

FromStrError

Error returned when decoding an AccountHash from a formatted string.

RemoveKeyFailure

Errors that can occur while removing a AccountHash from an account’s associated keys map.

SetThresholdFailure

Errors that can occur while changing action thresholds (i.e. the total Weights of signing AccountHashs required to perform various actions) on an account.

UpdateKeyFailure

Errors that can occur while updating the Weight of a AccountHash in an account’s associated keys map.

Constants

ACCOUNT_HASH_FORMATTED_STRING_PREFIX

The prefix applied to the hex-encoded AccountHash to produce a formatted string representation.

ACCOUNT_HASH_LENGTH

The length in bytes of a AccountHash.

WEIGHT_SERIALIZED_LENGTH

The number of bytes in a serialized Weight.