Struct casper_types::account::Account

pub struct Account { /* private fields */ }

Represents an Account in the global state.

Implementations

impl Account

pub fn new(
account_hash: AccountHash,
named_keys: NamedKeys,
main_purse: URef,
associated_keys: AssociatedKeys,
action_thresholds: ActionThresholds
) -> Self

Creates a new account.

pub fn create(
account: AccountHash,
named_keys: NamedKeys,
main_purse: URef
) -> Self

An Account constructor with presets for associated_keys and action_thresholds.

An account created with this method is valid and can be used as the target of a transaction. It will be created with an AssociatedKeys with a Weight of 1, and a default ActionThresholds.

pub fn extract_access_rights(&self) -> ContextAccessRights

Extracts the access rights from the named keys and main purse of the account.

pub fn named_keys_append(&mut self, keys: &mut NamedKeys)

Appends named keys to an account’s named_keys field.

pub fn named_keys(&self) -> &NamedKeys

Returns named keys.

pub fn named_keys_mut(&mut self) -> &mut NamedKeys

Returns a mutable reference to named keys.

pub fn account_hash(&self) -> AccountHash

Returns account hash.

pub fn main_purse(&self) -> URef

Returns main purse.

pub fn main_purse_add_only(&self) -> URef

Returns an AccessRights::ADD-only version of the main purse’s URef.

pub fn associated_keys(&self) -> &AssociatedKeys

Returns associated keys.

pub fn action_thresholds(&self) -> &ActionThresholds

Returns action thresholds.

pub fn add_associated_key(
&mut self,
account_hash: AccountHash,
weight: Weight
) -> Result<(), AddKeyFailure>

Adds an associated key to an account.

pub fn remove_associated_key(
&mut self,
account_hash: AccountHash
) -> Result<(), RemoveKeyFailure>

Removes an associated key from an account.

Verifies that removing the key will not cause the remaining weight to fall below any action thresholds.

pub fn update_associated_key(
&mut self,
account_hash: AccountHash,
weight: Weight
) -> Result<(), UpdateKeyFailure>

Updates an associated key.

Returns an error if the update would result in a violation of the key management thresholds.

pub fn set_action_threshold(
&mut self,
action_type: ActionType,
weight: Weight
) -> Result<(), SetThresholdFailure>

Sets new action threshold for a given action type for the account.

Returns an error if the new action threshold weight is greater than the total weight of the account’s associated keys.

pub fn can_set_threshold(
&self,
new_threshold: Weight
) -> Result<(), SetThresholdFailure>

Verifies if user can set action threshold.

pub fn can_authorize(&self, authorization_keys: &BTreeSet<AccountHash>) -> bool

Checks whether all authorization keys are associated with this account.

pub fn can_deploy_with(
&self,
authorization_keys: &BTreeSet<AccountHash>
) -> bool

Checks whether the sum of the weights of all authorization keys is greater or equal to deploy threshold.

pub fn can_manage_keys_with(
&self,
authorization_keys: &BTreeSet<AccountHash>
) -> bool

Checks whether the sum of the weights of all authorization keys is greater or equal to key management threshold.