Module casper_types::checksummed_hex

Checksummed hex encoding following an EIP-55-like scheme.

Constants

SMALL_BYTES_COUNT

The number of input bytes, at or below which decode will checksum-decode the output.

Functions

decode

Decodes a mixed-case hexadecimal string, verifying that it conforms to the checksum scheme similar to scheme in EIP-55.