Module casper_types::crypto

Cryptographic types and operations on them

Enums

Error

Cryptographic errors.

PublicKey

A public asymmetric key.

SecretKey

A secret or private asymmetric key.

Signature

A signature of given data.

Constants

ED25519_TAG

Tag for ed25519 variant.

SECP256K1_TAG

Tag for secp256k1 variant.

SYSTEM_ACCOUNT

Public key for system account.

SYSTEM_TAG

Tag for system variant.

Traits

AsymmetricType

Operations on asymmetric cryptographic type.

Functions

sign

Signs the given message using the given key pair.

verify

Verifies the signature of the given message against the given public key.