Struct casper_types::runtime_args::RuntimeArgs

pub struct RuntimeArgs(_);

Represents a collection of arguments passed to a smart contract.

Implementations

impl RuntimeArgs

pub fn new() -> RuntimeArgs

Create an empty RuntimeArgs instance.

pub fn try_new<F>(func: F) -> Result<RuntimeArgs, CLValueError> where
F: FnOnce(&mut RuntimeArgs) -> Result<(), CLValueError>,

A wrapper that lets you easily and safely create runtime arguments.

This method is useful when you have to construct a RuntimeArgs with multiple entries, but error handling at given call site would require to have a match statement for each RuntimeArgs::insert call. With this method you can use ? operator inside the closure and then handle single result. When try_block will be stabilized this method could be deprecated in favor of using those blocks.

pub fn get(&self, name: &str) -> Option<&CLValue>

Gets an argument by its name.

pub fn len(&self) -> usize

Gets the length of the collection.

pub fn is_empty(&self) -> bool

Returns true if the collection of arguments is empty.

pub fn insert<K, V>(&mut self, key: K, value: V) -> Result<(), CLValueError> where
K: Into<String>,
V: CLTyped + ToBytes,

Inserts a new named argument into the collection.

pub fn insert_cl_value<K>(&mut self, key: K, cl_value: CLValue) where
K: Into<String>,

Inserts a new named argument into the collection.

pub fn to_values(&self) -> Vec<&CLValue>

Returns all the values of the named args.

pub fn named_args(&self) -> impl Iterator<Item = &NamedArg>

Returns an iterator of references over all arguments in insertion order.

pub fn named_args_mut(&mut self) -> impl Iterator<Item = &mut NamedArg>

Returns an iterator of mutable references over all arguments in insertion order.

pub fn try_get_number(&self, name: &str) -> Result<U512, CLValueError>

Returns the numeric value of name arg from the runtime arguments or defaults to 0 if that arg doesn’t exist or is not an integer type.

Supported CLTypes for numeric conversions are U64, and U512.

Returns an error if parsing the arg fails.