Module casper_types::system::auction

Contains implementation of a Auction contract functionality.

Structs

Bid

An entry in the validator map.

Delegator

Represents a party delegating their stake to a validator (or “delegatee”)

EraInfo

Auction metadata. Intended to be recorded at each era.

SeigniorageRecipient

The seigniorage recipient details.

UnbondingPurse

Unbonding purse.

Enums

Error

Errors which can occur while executing the Auction contract.

SeigniorageAllocation

Information about a seigniorage allocation

Constants

ARG_AMOUNT

Named constant for amount.

ARG_AUCTION_DELAY

Named constant of auction_delay

ARG_DELEGATION_RATE

Named constant for delegation_rate.

ARG_DELEGATOR

Named constant for delegator.

ARG_DELEGATOR_PUBLIC_KEY

Named constant for delegator_public_key.

ARG_ERA_END_TIMESTAMP_MILLIS

Named constant for era_end_timestamp_millis;

ARG_ERA_ID

Named constant for era_id.

ARG_EVICTED_VALIDATORS

Named constant for evicted_validators;

ARG_GENESIS_VALIDATORS

Named constant for genesis_validators

ARG_LOCKED_FUNDS_PERIOD

Named constant for locked_funds_period

ARG_MINT_CONTRACT_PACKAGE_HASH

Named constant for mint_contract_package_hash

ARG_PUBLIC_KEY

Named constant for account_hash.

ARG_REWARD_FACTORS

Named constant for reward_factors.

ARG_UNBONDING_DELAY

Named constant for unbonding_delay

ARG_VALIDATOR

Named constant for validator.

ARG_VALIDATOR_KEYS

Named constant for validator_keys.

ARG_VALIDATOR_PUBLIC_KEY

Named constant for validator_public_key.

ARG_VALIDATOR_PUBLIC_KEYS

Named constant for validator_public_keys.

ARG_VALIDATOR_PURSE

Named constant for validator_purse.

ARG_VALIDATOR_SLOTS

Named constant for validator_slots argument.

AUCTION_DELAY_KEY

Amount of auction delay.

BLOCK_REWARD

We use one trillion as a block reward unit because it’s large enough to allow precise fractions, and small enough for many block rewards to fit into a u64.

DELEGATION_RATE_DENOMINATOR

Delegation rate is a fraction between 0-1. Validator sets the delegation rate in integer terms, which is then divided by the denominator to obtain the fraction.

ERA_END_TIMESTAMP_MILLIS_KEY

Storage for era-end timestamp.

ERA_ID_KEY

Storage for EraId.

INITIAL_ERA_END_TIMESTAMP_MILLIS

Initial value of era end timestamp.

INITIAL_ERA_ID

Initial value of era id we start at genesis.

LOCKED_FUNDS_PERIOD_KEY

Default lock period for new bid entries represented in eras.

METHOD_ACTIVATE_BID

Named constant for method activate_bid.

METHOD_ADD_BID

Named constant for method add_bid.

METHOD_DELEGATE

Named constant for method delegate.

METHOD_DISTRIBUTE

Named constant for method distribute.

METHOD_GET_ERA_VALIDATORS

Named constant for method get_era_validators.

METHOD_READ_ERA_ID

Named constant for method read_era_id.

METHOD_RUN_AUCTION

Named constant for method run_auction.

METHOD_SLASH

Named constant for method slash.

METHOD_UNDELEGATE

Named constant for method undelegate.

METHOD_WITHDRAW_BID

Named constant for method withdraw_bid.

SEIGNIORAGE_RECIPIENTS_SNAPSHOT_KEY

Storage for SeigniorageRecipientsSnapshot.

UNBONDING_DELAY_KEY

Unbonding delay expressed in eras.

VALIDATOR_SLOTS_KEY

Total validator slots allowed.

VESTING_SCHEDULE_LENGTH_MILLIS

Length of total vesting schedule expressed in days.

Traits

AccountProvider

Provider of an account related functionality.

Auction

Bonding auction contract interface

MintProvider

Provides an access to mint.

RuntimeProvider

Provider of runtime host functionality.

StorageProvider

Provides functionality of a contract storage.

Functions

auction_entry_points

Creates auction contract entry points.

era_validators_from_snapshot

Returns the era validators from a snapshot.

Type Definitions

Bids

Validators mapped to their bids.

DelegationRate

Representation of delegation rate of tokens. Range from 0..=100.

EraValidators

List of era validators

SeigniorageRecipients

Collection of seigniorage recipients.

SeigniorageRecipientsSnapshot

Snapshot of SeigniorageRecipients for a given era.

UnbondingPurses

Validators and delegators mapped to their unbonding purses.

ValidatorWeights

Weights of validators. “Weight” in this context means a sum of their stakes.