Module casper_types::system::handle_payment

Contains implementation of a Handle Payment contract functionality.

Enums

Error

Errors which can occur while executing the Handle Payment contract.

Constants

ARG_ACCOUNT

Named constant for source.

ARG_AMOUNT

Named constant for amount.

ARG_PURSE

Named constant for purse.

ARG_TARGET

Named constant for target.

CONTRACT_ACCESS_KEY

Storage for handle payment access key.

CONTRACT_HASH_KEY

Storage for handle payment contract hash.

METHOD_FINALIZE_PAYMENT

Named constant for method finalize_payment.

METHOD_GET_PAYMENT_PURSE

Named constant for method get_payment_purse.

METHOD_GET_REFUND_PURSE

Named constant for method get_refund_purse.

METHOD_SET_REFUND_PURSE

Named constant for method set_refund_purse.

PAYMENT_PURSE_KEY

The uref name where the Handle Payment accepts payment for computation on behalf of validators.

REFUND_PURSE_KEY

The uref name where the Handle Payment will refund unused payment back to the user. The uref this name corresponds to is set by the user.

Traits

HandlePayment

Handle payment functionality implementation.

MintProvider

Provides an access to mint.

RuntimeProvider

Provider of runtime host functionality.

Functions

handle_payment_entry_points

Creates handle payment contract entry points.