Enum casper_types::system::handle_payment::Error

#[repr(u8)]
pub enum Error {

    NotBonded,
    TooManyEventsInQueue,
    CannotUnbondLastValidator,
    SpreadTooHigh,
    MultipleRequests,
    BondTooSmall,
    BondTooLarge,
    UnbondTooLarge,
    BondTransferFailed,
    UnbondTransferFailed,
    TimeWentBackwards,
    StakesNotFound,
    PaymentPurseNotFound,
    PaymentPurseKeyUnexpectedType,
    PaymentPurseBalanceNotFound,
    BondingPurseNotFound,
    BondingPurseKeyUnexpectedType,
    RefundPurseKeyUnexpectedType,
    RewardsPurseNotFound,
    RewardsPurseKeyUnexpectedType,
    StakesKeyDeserializationFailed,
    StakesDeserializationFailed,
    SystemFunctionCalledByUserAccount,
    InsufficientPaymentForAmountSpent,
    FailedTransferToRewardsPurse,
    FailedTransferToAccountPurse,
    SetRefundPurseCalledOutsidePayment,
    GetBalance,
    PutKey,
    RemoveKey,
    Transfer,
    ArithmeticOverflow,
    RefundPurseIsPaymentPurse,
    // some variants omitted

}

Errors which can occur while executing the Handle Payment contract.

Variants

NotBonded

The given validator is not bonded.

assert_eq!(0, Error::NotBonded as u8);

TooManyEventsInQueue

There are too many bonding or unbonding attempts already enqueued to allow more.

assert_eq!(1, Error::TooManyEventsInQueue as u8);

CannotUnbondLastValidator

At least one validator must remain bonded.

assert_eq!(2, Error::CannotUnbondLastValidator as u8);

SpreadTooHigh

Failed to bond or unbond as this would have resulted in exceeding the maximum allowed difference between the largest and smallest stakes.

assert_eq!(3, Error::SpreadTooHigh as u8);

MultipleRequests

The given validator already has a bond or unbond attempt enqueued.

assert_eq!(4, Error::MultipleRequests as u8);

BondTooSmall

Attempted to bond with a stake which was too small.

assert_eq!(5, Error::BondTooSmall as u8);

BondTooLarge

Attempted to bond with a stake which was too large.

assert_eq!(6, Error::BondTooLarge as u8);

UnbondTooLarge

Attempted to unbond an amount which was too large.

assert_eq!(7, Error::UnbondTooLarge as u8);

BondTransferFailed

While bonding, the transfer from source purse to the Handle Payment internal purse failed.

assert_eq!(8, Error::BondTransferFailed as u8);

UnbondTransferFailed

While unbonding, the transfer from the Handle Payment internal purse to the destination purse failed.

assert_eq!(9, Error::UnbondTransferFailed as u8);

TimeWentBackwards

Internal error: a BlockTime was unexpectedly out of sequence.

assert_eq!(10, Error::TimeWentBackwards as u8);

StakesNotFound

Internal error: stakes were unexpectedly empty.

assert_eq!(11, Error::StakesNotFound as u8);

PaymentPurseNotFound

Internal error: the Handle Payment contract’s payment purse wasn’t found.

assert_eq!(12, Error::PaymentPurseNotFound as u8);

PaymentPurseKeyUnexpectedType

Internal error: the Handle Payment contract’s payment purse key was the wrong type.

assert_eq!(13, Error::PaymentPurseKeyUnexpectedType as u8);

PaymentPurseBalanceNotFound

Internal error: couldn’t retrieve the balance for the Handle Payment contract’s payment purse.

assert_eq!(14, Error::PaymentPurseBalanceNotFound as u8);

BondingPurseNotFound

Internal error: the Handle Payment contract’s bonding purse wasn’t found.

assert_eq!(15, Error::BondingPurseNotFound as u8);

BondingPurseKeyUnexpectedType

Internal error: the Handle Payment contract’s bonding purse key was the wrong type.

assert_eq!(16, Error::BondingPurseKeyUnexpectedType as u8);

RefundPurseKeyUnexpectedType

Internal error: the Handle Payment contract’s refund purse key was the wrong type.

assert_eq!(17, Error::RefundPurseKeyUnexpectedType as u8);

RewardsPurseNotFound

Internal error: the Handle Payment contract’s rewards purse wasn’t found.

assert_eq!(18, Error::RewardsPurseNotFound as u8);

RewardsPurseKeyUnexpectedType

Internal error: the Handle Payment contract’s rewards purse key was the wrong type.

assert_eq!(19, Error::RewardsPurseKeyUnexpectedType as u8);

StakesKeyDeserializationFailed

Internal error: failed to deserialize the stake’s key.

assert_eq!(20, Error::StakesKeyDeserializationFailed as u8);

StakesDeserializationFailed

Internal error: failed to deserialize the stake’s balance.

assert_eq!(21, Error::StakesDeserializationFailed as u8);

SystemFunctionCalledByUserAccount

The invoked Handle Payment function can only be called by system contracts, but was called by a user contract.

assert_eq!(22, Error::SystemFunctionCalledByUserAccount as u8);

InsufficientPaymentForAmountSpent

Internal error: while finalizing payment, the amount spent exceeded the amount available.

assert_eq!(23, Error::InsufficientPaymentForAmountSpent as u8);

FailedTransferToRewardsPurse

Internal error: while finalizing payment, failed to pay the validators (the transfer from the Handle Payment contract’s payment purse to rewards purse failed).

assert_eq!(24, Error::FailedTransferToRewardsPurse as u8);

FailedTransferToAccountPurse

Internal error: while finalizing payment, failed to refund the caller’s purse (the transfer from the Handle Payment contract’s payment purse to refund purse or account’s main purse failed).

assert_eq!(25, Error::FailedTransferToAccountPurse as u8);

SetRefundPurseCalledOutsidePayment

Handle Payment contract’s “set_refund_purse” method can only be called by the payment code of a deploy, but was called by the session code.

assert_eq!(26, Error::SetRefundPurseCalledOutsidePayment as u8);

GetBalance

Raised when the system is unable to determine purse balance.

assert_eq!(27, Error::GetBalance as u8);

PutKey

Raised when the system is unable to put named key.

assert_eq!(28, Error::PutKey as u8);

RemoveKey

Raised when the system is unable to remove given named key.

assert_eq!(29, Error::RemoveKey as u8);

Transfer

Failed to transfer funds.

assert_eq!(30, Error::Transfer as u8);

ArithmeticOverflow

An arithmetic overflow occurred

assert_eq!(31, Error::ArithmeticOverflow as u8);

RefundPurseIsPaymentPurse

Refund purse is a payment purse.