Trait casper_types::system::handle_payment::MintProvider

pub trait MintProvider {
    fn transfer_purse_to_account(
        &mut self, 
        source: URef, 
        target: AccountHash, 
        amount: U512
    ) -> Result<TransferredTo, Error>;

    fn transfer_purse_to_purse(
        &mut self, 
        source: URef, 
        target: URef, 
        amount: U512
    ) -> Result<(), Error>;

    fn balance(&mut self, purse: URef) -> Result<Option<U512>, Error>;
}

Provides an access to mint.

Required methods

fn transfer_purse_to_account(
&mut self,
source: URef,
target: AccountHash,
amount: U512
) -> Result<TransferredTo, Error>

Transfer amount from source purse to a target account.

fn transfer_purse_to_purse(
&mut self,
source: URef,
target: URef,
amount: U512
) -> Result<(), Error>

Transfer amount from source purse to a target purse.

fn balance(&mut self, purse: URef) -> Result<Option<U512>, Error>

Checks balance of a purse. Returns None if given purse does not exist.